Code cleanup to allign with coding rules.
