+++ protocol/src/common/protocol.ts
@@ -64,7 +64,10 @@ import {
 import {
 	OnTypeRenameClientCapabilities, OnTypeRenameRanges, OnTypeRenameOptions, OnTypeRenameParams, OnTypeRenameRegistrationOptions, OnTypeRenameRequest
 } from './protocol.onTypeRename';
-import { FileOperationClientCapabilities } from './protocol.window.fileOperations';
+
+import {
+	FileOperationOptions, FileOperationClientCapabilities, FileOperationRegistrationOptions, WillCreateFilesRequest, CreateFilesParams, FileCreate, FileRename, FileDelete
+} from './protocol.window.fileOperations';
 
 // @ts-ignore: to avoid inlining LocationLink as dynamic import
 let __noDynamicImport: LocationLink | undefined;
@@ -1635,52 +1638,6 @@ export namespace WatchKind {
 	export const Delete = 4;
 }
 
-//---- User file events ----
-
-/**
- * Options for notifications/requests for user operations on files.
- */
-export interface FileOperationOptions {
-	/**
-	* The server is interested in didCreateFiles notifications.
-	*/
-	didCreate?: FileOperationRegistrationOptions;
-	/**
-	* The server is interested in willCreateFiles requests.
-	*/
-	willCreate?: FileOperationRegistrationOptions;
-	/**
-	* The server is interested in didRenameFiles notifications.
-	*/
-	didRename?: FileOperationRegistrationOptions;
-	/**
-	* The server is interested in willRenameFiles requests.
-	*/
-	willRename?: FileOperationRegistrationOptions;
-	/**
-	* The server is interested in didDeleteFiles file notifications.
-	*/
-	didDelete?: FileOperationRegistrationOptions;
-	/**
-	* The server is interested in willDeleteFiles file requests.
-	*/
-	willDelete?: FileOperationRegistrationOptions;
-}
-
-export interface FileOperationRegistrationOptions {
-	/**
-	 * The glob pattern to match. Glob patterns can have the following syntax:
-	 * - `*` to match one or more characters in a path segment
-	 * - `?` to match on one character in a path segment
-	 * - `**` to match any number of path segments, including none
-	 * - `{}` to group conditions (e.g. `**​/*.{ts,js}` matches all TypeScript and JavaScript files)
-	 * - `[]` to declare a range of characters to match in a path segment (e.g., `example.[0-9]` to match on `example.0`, `example.1`, …)
-	 * - `[!...]` to negate a range of characters to match in a path segment (e.g., `example.[!0-9]` to match on `example.a`, `example.b`, but not `example.0`)
-	 * - `/` suffix to match only folders (e.g. `**{/,*.dart}` matches all Dart files and all folders)
-	 */
-	globPattern: string;
-}
-
 //---- Diagnostic notification ----
 
 /**
@@ -3221,7 +3178,9 @@ export {
 	// Show document
 	ShowDocumentParams, ShowDocumentRequest, ShowDocumentResult, ShowDocumentClientCapabilities,
 	// On Type rename
-	OnTypeRenameClientCapabilities, OnTypeRenameRanges, OnTypeRenameOptions, OnTypeRenameParams, OnTypeRenameRegistrationOptions, OnTypeRenameRequest
+	OnTypeRenameClientCapabilities, OnTypeRenameRanges, OnTypeRenameOptions, OnTypeRenameParams, OnTypeRenameRegistrationOptions, OnTypeRenameRequest,
+	// File operations
+	FileOperationClientCapabilities, FileOperationOptions, FileOperationRegistrationOptions, CreateFilesParams, WillCreateFilesRequest, FileCreate, FileRename, FileDelete
 };
 
 // To be backwards compatible
