+++ protocol/src/common/protocol.window.fileOperations.ts
@@ -6,7 +6,50 @@
 import { RequestHandler } from 'vscode-jsonrpc';
 import { WorkspaceEdit } from 'vscode-languageserver-types';
 import { ProtocolRequestType } from './messages';
-import { FileOperationRegistrationOptions } from './protocol';
+
+/**
+ * Options for notifications/requests for user operations on files.
+ */
+export interface FileOperationOptions {
+	/**
+	* The server is interested in didCreateFiles notifications.
+	*/
+	didCreate?: FileOperationRegistrationOptions;
+	/**
+	* The server is interested in willCreateFiles requests.
+	*/
+	willCreate?: FileOperationRegistrationOptions;
+	/**
+	* The server is interested in didRenameFiles notifications.
+	*/
+	didRename?: FileOperationRegistrationOptions;
+	/**
+	* The server is interested in willRenameFiles requests.
+	*/
+	willRename?: FileOperationRegistrationOptions;
+	/**
+	* The server is interested in didDeleteFiles file notifications.
+	*/
+	didDelete?: FileOperationRegistrationOptions;
+	/**
+	* The server is interested in willDeleteFiles file requests.
+	*/
+	willDelete?: FileOperationRegistrationOptions;
+}
+
+export interface FileOperationRegistrationOptions {
+	/**
+	 * The glob pattern to match. Glob patterns can have the following syntax:
+	 * - `*` to match one or more characters in a path segment
+	 * - `?` to match on one character in a path segment
+	 * - `**` to match any number of path segments, including none
+	 * - `{}` to group conditions (e.g. `**​/*.{ts,js}` matches all TypeScript and JavaScript files)
+	 * - `[]` to declare a range of characters to match in a path segment (e.g., `example.[0-9]` to match on `example.0`, `example.1`, …)
+	 * - `[!...]` to negate a range of characters to match in a path segment (e.g., `example.[!0-9]` to match on `example.a`, `example.b`, but not `example.0`)
+	 * - `/` suffix to match only folders (e.g. `**{/,*.dart}` matches all Dart files and all folders)
+	 */
+	globPattern: string;
+}
 
 /**
  * Capabilities relating to events from file operations by the user in the client.
@@ -112,4 +155,5 @@ export interface FileDelete {
 export namespace WillCreateFilesRequest {
 	export const method: 'window/willCreateFiles' = 'window/willCreateFiles';
 	export const type = new ProtocolRequestType<CreateFilesParams, WorkspaceEdit | null, never, void, FileOperationRegistrationOptions>(method);
+	export type HandlerSignature = RequestHandler<CreateFilesParams, WorkspaceEdit | undefined | null, void>;
 }
